Change type of attributes

+++ app/src/main/java/com/marwaeltayeb/weatherforecast/model/Main.kt
@@ -14,11 +14,11 @@ data class Main(
     val pressure: Int,
 
     @SerializedName("temp")
-    val temp: Int,
+    val temp: Double,
 
     @SerializedName("temp_max")
-    val tempMax: Int,
+    val tempMax: Double,
 
     @SerializedName("temp_min")
-    val tempMin: Int
+    val tempMin: Double
 )
